@@ -4,19 +4,19 @@
   "description": "",
   "main": "index.js",
   "dependencies": {
-    "@typescript-eslint/parser": "^4.31.0",
-    "eslint-config-prettier": "^8.3.0",
-    "eslint-plugin-promise": "^5.1.0",
+    "@typescript-eslint/parser": "^5.62.0",
+    "eslint-config-prettier": "^8.10.0",
+    "eslint-plugin-promise": "^6.1.1",
     "typescript": "^4.9.5"
   },
   "devDependencies": {
-    "@babel/core": "^7.14.8",
-    "@typescript-eslint/eslint-plugin": "^4.28.4",
+    "@babel/core": "^7.22.20",
+    "@typescript-eslint/eslint-plugin": "^5.62.0",
     "babel-eslint": "^10.1.0",
-    "eslint": "^7.32.0",
-    "eslint-plugin-import": "^2.23.4",
-    "eslint-plugin-react": "^7.24.0",
-    "eslint-plugin-react-hooks": "^4.2.0",
+    "eslint": "^8.56.0",
+    "eslint-plugin-import": "^2.29.1",
+    "eslint-plugin-react": "^7.33.2",
+    "eslint-plugin-react-hooks": "^4.6.0",
     "prettier": "2.8.8"
   },
   "scripts": {
@@ -28,8 +28,8 @@
     "install:frontend": "cd frontend && npm install",
     "install:server": "cd server && npm install",
     "postinstall": "npm run install:frontend && npm run install:server",
-    "quiet-lint": "npx eslint . --ext .js --ext .jsx --ext .ts --ext .tsx --fix --quiet",
-    "lint": "npx eslint . --ext .js --ext .jsx --ext .ts --ext .tsx --fix",
+    "quiet-lint": "npx eslint . --ext .js,.jsx,.ts,.tsx --fix --quiet",
+    "lint": "npx eslint . --ignore-path .eslintignore --ext .js,.jsx,.ts,.tsx --fix",
     "format": "npx prettier --write \"**/*.{ts,js,tsx,jsx,css}\"",
     "format:check": "npx prettier --check \"**/*.{ts,js,tsx,jsx,css}\"",
     "type-check": "cd frontend && npm run type-check && cd ../server && npm run type-check",
