@@ -1,10 +1,18 @@
 import React, { useCallback, useEffect, useState } from 'react';
 import cn from 'classnames';
 import moment from 'moment';
-import { useFormContext } from 'react-hook-form';
+import { useFormContext, UseFormRegister } from 'react-hook-form';
 import styles from './employeemodal.module.css';
 import { Input, Label } from '../FormElements/FormElements';
 import { WeekProvider, useWeek } from './WeekContext';
+
+type FormData = {
+  availability: {
+    startTime: string;
+    endTime: string;
+    days: string[];
+  }[];
+};
 
 type AvailabilityInputProps = {
   index: number;
@@ -13,12 +21,12 @@ type AvailabilityInputProps = {
   hide: boolean;
 };
 
-const AvailabilityInput = ({
+const AvailabilityInput: React.FC<AvailabilityInputProps> = ({
   index,
   existingTimeRange,
   existingDayArray,
   hide,
-}: AvailabilityInputProps) => {
+}) => {
   const {
     selectDay,
     deselectDay,
@@ -26,8 +34,12 @@ const AvailabilityInput = ({
     isDaySelectedByInstance,
     getSelectedDays,
   } = useWeek();
-  const { register, setValue, getValues, formState } = useFormContext();
-  const { errors } = formState;
+  const {
+    register,
+    setValue,
+    getValues,
+    formState: { errors },
+  } = useFormContext<FormData>();
   const dayLabels = {
     Mon: 'M',
     Tue: 'T',
@@ -35,11 +47,10 @@ const AvailabilityInput = ({
     Thu: 'T',
     Fri: 'F',
   };
-  const [existingTime, setExisingTime] = useState<string[]>();
-  // All data for this AvailabilityInput instance will be saved in the form's
-  // data in an array 'availability' at index 'index'
-  const instance = `availability[${index}]`;
+  const [existingTime, setExistingTime] = useState<string[]>();
+  const instance = `availability.${index}` as const;
   const days = getSelectedDays(index);
+
   const handleClick = (day: string) => {
     if (isDaySelectedByInstance(day, index)) {
       deselectDay(day);
@@ -48,45 +59,30 @@ const AvailabilityInput = ({
     }
   };
 
-  const prefillDays = () => {
+  const prefillDays = useCallback(() => {
     existingDayArray?.forEach((day) => {
       selectDay(day, index);
     });
-  };
+  }, [existingDayArray, index, selectDay]);
 
-  const prefillTimeRange = () => {
+  const prefillTimeRange = useCallback(() => {
     if (existingTimeRange) {
-      // extract start and end times
       let [startTime, endTime] = existingTimeRange.split('-');
       startTime = formatTime(startTime);
       endTime = formatTime(endTime);
-      setExisingTime([startTime, endTime]);
+      setExistingTime([startTime, endTime]);
     }
-  };
+  }, [existingTimeRange]);
 
   useEffect(() => {
-    // Prefill days and time range once
     prefillDays();
     prefillTimeRange();
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, []);
+  }, [prefillDays, prefillTimeRange]);
 
   useEffect(() => {
-    // Register day selector as custom form input.
-    // Not putting error message here since there is no default behavior to override
-    register(`${instance}.days`, {
-      required: !hide,
-      validate: () => (hide ? true : days.length > 0),
-    });
-  }, [instance, register, days, hide]);
-
-  useEffect(() => {
-    // When selected days changes, update days value
     setValue(`${instance}.days`, days);
   }, [instance, days, setValue]);
 
-  // formats a time string to be suitable to pass into defaultValue
-  // of the Input component
   const formatTime = (time: string): string =>
     moment(time, 'ha').format('HH:mm');
 
@@ -96,40 +92,34 @@ const AvailabilityInput = ({
         <Label htmlFor={`${instance}.startTime`}>Start Time</Label>
         <Input
           id={`${instance}.startTime`}
-          name={`${instance}.startTime`}
           type="time"
           className={styles.timeInput}
           defaultValue={existingTime?.[0]}
-          ref={register({ required: !hide })}
+          {...register(`${instance}.startTime` as const, { required: !hide })}
         />
-        {errors.availability &&
-          errors.availability[index] &&
-          errors.availability[index].startTime && (
-            <p className={styles.error}>Please enter a valid start time</p>
-          )}
+        {errors.availability?.[index]?.startTime && (
+          <p className={styles.error}>Please enter a valid start time</p>
+        )}
       </div>
       <p className={styles.toText}>to</p>
       <div className={styles.timeFlexbox}>
         <Label htmlFor={`${instance}.endTime`}>End Time</Label>
         <Input
           id={`${instance}.endTime`}
-          name={`${instance}.endTime`}
           type="time"
           className={styles.timeInput}
           defaultValue={existingTime?.[1]}
-          ref={register({
+          {...register(`${instance}.endTime` as const, {
             required: !hide,
-            validate: (endTime) => {
-              const startTime = getValues(`${instance}.startTime`);
+            validate: (endTime: string) => {
+              const startTime = getValues(`${instance}.startTime` as const);
               return hide ? true : startTime < endTime;
             },
           })}
         />
-        {errors.availability &&
-          errors.availability[index] &&
-          errors.availability[index].endTime && (
-            <p className={styles.error}>Please enter a valid end time</p>
-          )}
+        {errors.availability?.[index]?.endTime && (
+          <p className={styles.error}>Please enter a valid end time</p>
+        )}
       </div>
       <p className={styles.repeatText}>Repeat on</p>
       <div className={styles.timeFlexbox}>
@@ -137,7 +127,6 @@ const AvailabilityInput = ({
           {Object.entries(dayLabels).map(([day, label]) => (
             <Input
               key={day}
-              name={`${instance}.days`}
               type="button"
               value={label}
               className={cn(styles.day, {
@@ -147,13 +136,11 @@ const AvailabilityInput = ({
             />
           ))}
         </div>
-        {errors.availability &&
-          errors.availability[index] &&
-          errors.availability[index].days && (
-            <p className={cn(styles.error, styles.dayError)}>
-              Please select at least one day
-            </p>
-          )}
+        {errors.availability?.[index]?.days && (
+          <p className={cn(styles.error, styles.dayError)}>
+            Please select at least one day
+          </p>
+        )}
       </div>
     </div>
   );
@@ -164,54 +151,41 @@ type WorkingHoursProps = {
   hide: boolean;
 };
 
-const WorkingHours = ({ existingAvailability, hide }: WorkingHoursProps) => {
+const WorkingHours: React.FC<WorkingHoursProps> = ({
+  existingAvailability,
+  hide,
+}) => {
   const [numAvailability, setNumAvailability] = useState(
     existingAvailability ? 0 : 1
   );
-  const [availabilityArray, setAvailabilityArray] = useState<any[]>([]);
+  const [availabilityArray, setAvailabilityArray] = useState<
+    [string, string[]][]
+  >([]);
 
   const addAvailabilityInput = () => setNumAvailability((n) => n + 1);
 
-  // returns a map with time ranges as keys and repeating days as values
   const getAvailabilityMap = useCallback((): Map<string, string[]> => {
     const availabilityMap = new Map();
     existingAvailability?.forEach((availability) => {
       const [day, timeRange] = availability;
-      const dayArray: string[] | undefined = availabilityMap.get(timeRange);
-      if (dayArray) {
-        // push day onto dayArray
-        dayArray.push(day);
-        availabilityMap.set(timeRange, dayArray);
-      } else {
-        // create new array
-        availabilityMap.set(timeRange, [day]);
-      }
+      const dayArray = availabilityMap.get(timeRange) || [];
+      dayArray.push(day);
+      availabilityMap.set(timeRange, dayArray);
     });
     return availabilityMap;
   }, [existingAvailability]);
 
-  // transforms the availability map into an array
-  const availabilityMapToArray = (map: Map<string, string[]>) => {
-    const timeRangeArray = Array.from(map.keys());
-    const dayArray = Array.from(map.values());
-    if (timeRangeArray.length === dayArray.length) {
-      // Each element of newAvailabilityArray is a tuple of (string, string[])
-      const newAvailabilityArray: any[] = [];
-      for (let i = 0; i < timeRangeArray.length; i += 1) {
-        const pair = [timeRangeArray[i], dayArray[i]];
-        newAvailabilityArray.push(pair);
-      }
-      setAvailabilityArray(newAvailabilityArray);
-    } else {
-      console.log(
-        'Error: Sizes of time and day arrays are not equal. This error should be impossible.'
-      );
-    }
-  };
+  const availabilityMapToArray = useCallback((map: Map<string, string[]>) => {
+    const newAvailabilityArray: [string, string[]][] = Array.from(
+      map,
+      ([timeRange, dayArray]) => [timeRange, dayArray]
+    );
+    setAvailabilityArray(newAvailabilityArray);
+  }, []);
 
   useEffect(() => {
     availabilityMapToArray(getAvailabilityMap());
-  }, [getAvailabilityMap]);
+  }, [getAvailabilityMap, availabilityMapToArray]);
 
   return (
     <div className={cn(styles.workingHours, { [styles.hidden]: hide })}>
@@ -227,12 +201,12 @@ const WorkingHours = ({ existingAvailability, hide }: WorkingHoursProps) => {
                 hide={hide}
               />
             ))
-          : [...new Array(numAvailability)].map((_, index) => (
+          : [...Array(numAvailability)].map((_, index) => (
               <AvailabilityInput key={index} index={index} hide={hide} />
             ))}
       </WeekProvider>
       <button
-        type={'button'}
+        type="button"
         className={styles.addAvailabilityInput}
         onClick={addAvailabilityInput}
       >
