@@ -1,5 +1,5 @@
 import React, { useState } from 'react';
-import { useForm } from 'react-hook-form';
+import { useForm, SubmitHandler } from 'react-hook-form';
 import cn from 'classnames';
 import { Button, Input, Label } from '../FormElements/FormElements';
 import styles from './ridermodal.module.css';
@@ -12,26 +12,43 @@ type ModalFormProps = {
   rider?: Rider;
 };
 
-const RiderModalInfo = ({
+type FormData = {
+  firstName: string;
+  lastName: string;
+  netid: string;
+  phoneNumber: string;
+  needs: string;
+  address: string;
+  joinDate: string;
+  endDate: string;
+  otherNeeds?: string;
+};
+
+const RiderModalInfo: React.FC<ModalFormProps> = ({
   onSubmit,
   setIsOpen,
   setFormData,
   rider,
-}: ModalFormProps) => {
-  const { register, formState, handleSubmit, getValues } = useForm({
+}) => {
+  const {
+    register,
+    formState: { errors },
+    handleSubmit,
+    getValues,
+  } = useForm<FormData>({
     defaultValues: {
       firstName: rider?.firstName ?? '',
       lastName: rider?.lastName ?? '',
       netid: rider?.email.split('@')[0] ?? '',
       phoneNumber: rider?.phoneNumber ?? '',
-      needs: rider?.accessibility ?? '', // if no needs, default is undefined
+      needs: rider?.accessibility ?? '',
       address: rider?.address ?? '',
       joinDate: rider?.joinDate ?? '',
       endDate: rider?.endDate ?? '',
     },
   });
-  const { errors } = formState;
-  const beforeSubmit = ({
+
+  const beforeSubmit: SubmitHandler<FormData> = ({
     firstName,
     lastName,
     netid,
@@ -40,7 +57,7 @@ const RiderModalInfo = ({
     address,
     joinDate,
     endDate,
-  }: ObjectType) => {
+  }) => {
     const email = netid ? `${netid}@cornell.edu` : undefined;
     const accessibility = needs;
     onSubmit({
@@ -64,6 +81,7 @@ const RiderModalInfo = ({
   const isEditing = rider !== undefined;
   const isStudentEditing = isEditing && localUserType === 'Rider';
   const [needsOption, setNeedsOption] = useState('');
+
   return (
     <form onSubmit={handleSubmit(beforeSubmit)} className={styles.form}>
       <div className={cn(styles.inputContainer, styles.rideTime)}>
@@ -73,9 +91,8 @@ const RiderModalInfo = ({
           </Label>
           <Input
             id="firstName"
-            name="firstName"
+            {...register('firstName', { required: true })}
             type="text"
-            ref={register({ required: true })}
             aria-required="true"
             className={styles.firstRow}
           />
@@ -87,9 +104,8 @@ const RiderModalInfo = ({
           </Label>
           <Input
             id="lastName"
-            name="lastName"
+            {...register('lastName', { required: true })}
             type="text"
-            ref={register({ required: true })}
             className={styles.firstRow}
             aria-required="true"
           />
@@ -103,9 +119,11 @@ const RiderModalInfo = ({
           </Label>
           <Input
             id="netid"
-            name="netid"
+            {...register('netid', {
+              required: true,
+              pattern: /^[a-zA-Z]+[0-9]+$/,
+            })}
             type="text"
-            ref={register({ required: true, pattern: /^[a-zA-Z]+[0-9]+$/ })}
             disabled={isStudentEditing}
             className={styles.firstRow}
             aria-required="true"
@@ -120,9 +138,11 @@ const RiderModalInfo = ({
           </Label>
           <Input
             id="phoneNumber"
-            name="phoneNumber"
+            {...register('phoneNumber', {
+              required: true,
+              pattern: /^[0-9]{10}$/,
+            })}
             type="tel"
-            ref={register({ required: true, pattern: /^[0-9]{10}$/ })}
             className={styles.firstRow}
             aria-required="true"
           />
@@ -136,28 +156,23 @@ const RiderModalInfo = ({
           </Label>
           <select
             id="needs"
-            name="needs"
+            {...register('needs', { required: true })}
             aria-required="true"
-            ref={register({ required: true })}
             onChange={(e) => setNeedsOption(e.target.value)}
           >
-            {Object.values(Accessibility).map((value, index) => {
-              return (
-                <option key={index} value={value}>
-                  {value}
-                </option>
-              );
-            })}
+            {Object.values(Accessibility).map((value, index) => (
+              <option key={index} value={value}>
+                {value}
+              </option>
+            ))}
           </select>
-          {needsOption == 'Other' ? (
+          {needsOption === 'Other' && (
             <Input
               id="otherNeeds"
-              name="otherNeeds"
+              {...register('otherNeeds')}
               type="text"
               placeholder="Please Specify Needs"
-            ></Input>
-          ) : (
-            <></>
+            />
           )}
           {errors.needs?.type === 'validate' && (
             <p className={styles.error}>
@@ -172,13 +187,12 @@ const RiderModalInfo = ({
           </Label>
           <Input
             id="address"
-            name="address"
-            type="text"
-            aria-required="true"
-            ref={register({
+            {...register('address', {
               required: true,
               pattern: /^[a-zA-Z0-9\s,.'-]{3,}$/,
             })}
+            type="text"
+            aria-required="true"
           />
           {errors.address && (
             <p className={styles.error}>Please enter an address</p>
@@ -193,10 +207,9 @@ const RiderModalInfo = ({
               </Label>
               <Input
                 id="joinDate"
+                {...register('joinDate', { required: true })}
                 type="date"
-                name="joinDate"
                 aria-required="true"
-                ref={register({ required: true })}
                 disabled={isStudentEditing}
                 className={styles.riderDate}
               />
@@ -211,16 +224,15 @@ const RiderModalInfo = ({
               </Label>
               <Input
                 id="endDate"
-                type="date"
-                name="endDate"
-                aria-required="true"
-                ref={register({
+                {...register('endDate', {
                   required: true,
                   validate: (endDate) => {
                     const joinDate = getValues('joinDate');
                     return joinDate < endDate;
                   },
                 })}
+                type="date"
+                aria-required="true"
                 disabled={isStudentEditing}
                 className={styles.riderDate}
               />
