@@ -1,12 +1,12 @@
 import express from 'express';
 import jwt from 'jsonwebtoken';
-import { ModelType } from 'dynamoose/dist/General';
-import { Document } from 'dynamoose/dist/Document';
 import { Rider } from '../models/rider';
 import { Admin } from '../models/admin';
 import { Driver } from '../models/driver';
 import { OAuth2Client } from 'google-auth-library';
 import { oauthValues } from '../config';
+import { ModelType } from 'dynamoose/dist/General';
+import { Item } from 'dynamoose/dist/Item';
 
 const router = express.Router();
 
@@ -22,7 +22,7 @@ const audience = [
 ];
 
 function getModel(table: string) {
-  const tableToModel: { [table: string]: ModelType<Document> } = {
+  const tableToModel: { [table: string]: ModelType<Item> } = {
     Riders: Rider,
     Drivers: Driver,
     Admins: Admin,
@@ -36,7 +36,7 @@ function getUserType(table: string) {
 
 function findUserAndSendToken(
   res: express.Response,
-  model: ModelType<Document>,
+  model: ModelType<Item>,
   table: string,
   email: string
 ) {
